migrate switchable runtime tests
