+++ test/e2e/switchable-runtime/app/node-rsc-ssg/page.server.js
@@ -1,5 +1,5 @@
-import Runtime from '../utils/runtime'
-import Time from '../utils/time'
+import Runtime from '../../utils/runtime'
+import Time from '../../utils/time'
 
 export default function Page({ type }) {
   return (
