+++ test/e2e/switchable-runtime/pages/app.js.txt
@@ -1,6 +1,6 @@
 export default function App({ Component, pageProps }) {
   return (
-    <div className="app-client-root" data-title={Component.title || ''}>
+    <div className="app-client-root">
       <Component {...pageProps} />
     </div>
   )
