+++ test/e2e/switchable-runtime/pages/edge.js
