+++ test/e2e/switchable-runtime/pages/static.js
