+++ test/integration/react-18-invalid-config/index.test.js
@@ -5,38 +5,12 @@ import { join } from 'path'
 import {
   File,
   nextBuild,
-  runDevSuite,
-  runProdSuite,
-  fetchViaHTTP,
 } from 'next-test-utils'
 
 const appDir = __dirname
 const nodeArgs = ['-r', join(appDir, '../../lib/react-channel-require-hook.js')]
 const reactDomPackagePah = join(appDir, 'node_modules/react-dom')
 const nextConfig = new File(join(appDir, 'next.config.js'))
-const documentPage = new File(join(appDir, 'pages/_document.js'))
-const indexPage = join(appDir, 'pages/index.js')
-const indexServerPage = join(appDir, 'pages/index.server.js')
-
-const documentWithGip = `
-import { Html, Head, Main, NextScript } from 'next/document'
-
-export default function Document() {
-  return (
-    <Html>
-      <Head />
-      <body>
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  )
-}
-
-Document.getInitialProps = (ctx) => {
-  return ctx.defaultGetInitialProps(ctx)
-}
-`
 
 function writeNextConfig(config) {
   const content = `
@@ -91,36 +65,3 @@ describe('React 17 with React 18 config', () => {
     expect(code).toBe(1)
   })
 })
-
-const documentSuite = {
-  runTests: (context, env) => {
-    if (env === 'dev') {
-      it('should error when custom _document has getInitialProps method', async () => {
-        const res = await fetchViaHTTP(context.appPort, '/')
-        expect(res.status).toBe(500)
-      })
-    } else {
-      it('should failed building', async () => {
-        expect(context.code).toBe(1)
-      })
-    }
-  },
-  beforeAll: async () => {
-    writeNextConfig({
-      serverComponents: true,
-    })
-    documentPage.write(documentWithGip)
-    await fs.rename(indexPage, indexServerPage)
-  },
-  afterAll: async () => {
-    documentPage.delete()
-    nextConfig.restore()
-    await fs.rename(indexServerPage, indexPage)
-  },
-  env: {
-    __NEXT_REACT_CHANNEL: 'exp',
-  },
-}
-
-runDevSuite('Invalid custom document with gip', appDir, documentSuite)
-runProdSuite('Invalid custom document with gip', appDir, documentSuite)
