--- test/integration/react-server-components/switchable-runtime/package.json
@@ -1,8 +1,0 @@
-{
-  "private": true,
-  "scripts": {
-    "dev": "yarn next dev",
-    "build": "yarn next build",
-    "start": "yarn next start"
-  }
-}
